CORE-32927 Renamed createCookie to createComponentNamespacedCookieJar.
