+++ src/core/createComponentNamespacedCookieJar.js
